Add binary bitmap font support | GUI support for keyboard
Add new input box
Completely moved from GNU to NASM
Updated NINA and VBE for binary bitmap drawing

+++ core/globals.cpp
@@ -1,5 +1,5 @@
 #include <core/globals.h>
 
 char Buffer[32];
-uint32_t timerTicks = 0;
+uint64_t timerTicks = 0;
